- Move the Scheduler time tick broadcast receiver into Aware service (as it is a foreground service) to test if it will perform better.
