+++ aware-core/src/main/java/com/aware/Applications.java
@@ -6,13 +6,10 @@ import android.accessibilityservice.AccessibilityServiceInfo;
 import android.app.ActivityManager;
 import android.app.ActivityManager.ProcessErrorStateInfo;
 import android.app.ActivityManager.RunningAppProcessInfo;
-import android.app.AlarmManager;
 import android.app.IntentService;
 import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
-import android.app.Service;
-import android.content.BroadcastReceiver;
 import android.content.ContentValues;
 import android.content.Context;
 import android.content.Intent;
@@ -28,13 +25,10 @@ import android.database.sqlite.SQLiteException;
 import android.net.Uri;
 import android.os.Build;
 import android.provider.Settings;
-import android.support.annotation.IntDef;
-import android.support.v4.accessibilityservice.AccessibilityServiceInfoCompat;
 import android.support.v4.app.NotificationCompat;
 import android.support.v4.content.WakefulBroadcastReceiver;
 import android.support.v4.view.accessibility.AccessibilityEventCompat;
 import android.support.v4.view.accessibility.AccessibilityManagerCompat;
-import android.text.TextUtils;
 import android.util.Log;
 import android.view.accessibility.AccessibilityEvent;
 import android.view.accessibility.AccessibilityManager;
@@ -294,10 +288,6 @@ public class Applications extends AccessibilityService {
         webservices.addAction(Aware.ACTION_AWARE_CLEAR_DATA);
         registerReceiver(awareMonitor, webservices);
 
-        IntentFilter scheduler = new IntentFilter();
-        scheduler.addAction(Intent.ACTION_TIME_TICK);
-        registerReceiver(schedulerTicker, scheduler);
-
         if (Aware.getSetting(getApplicationContext(), Aware_Preferences.FOREGROUND_PRIORITY).equals("true")) {
             sendBroadcast(new Intent(Aware.ACTION_AWARE_PRIORITY_FOREGROUND));
         }
@@ -353,7 +343,6 @@ public class Applications extends AccessibilityService {
         if (Aware.getSetting(getApplicationContext(), Applications.STATUS_AWARE_ACCESSIBILITY).equals("true")) {
             try {
                 if (awareMonitor != null) unregisterReceiver(awareMonitor);
-                if (schedulerTicker != null) unregisterReceiver(schedulerTicker);
             } catch (IllegalArgumentException e) {
             }
         }
@@ -369,7 +358,6 @@ public class Applications extends AccessibilityService {
         if (Aware.getSetting(getApplicationContext(), Applications.STATUS_AWARE_ACCESSIBILITY).equals("true")) {
             try {
                 if (awareMonitor != null) unregisterReceiver(awareMonitor);
-                if (schedulerTicker != null) unregisterReceiver(schedulerTicker);
             } catch (IllegalArgumentException e) {
             }
         }
@@ -479,19 +467,6 @@ public class Applications extends AccessibilityService {
         return true;
     }
 
-    private final SchedulerTicker schedulerTicker = new SchedulerTicker();
-    public class SchedulerTicker extends BroadcastReceiver {
-        @Override
-        public void onReceive(Context context, Intent intent) {
-            //Executed every 1-minute. OS will send this tickle automatically
-            if (intent.getAction().equals(Intent.ACTION_TIME_TICK)) {
-                Intent scheduler = new Intent(context, Scheduler.class);
-                scheduler.setAction(Scheduler.ACTION_AWARE_SCHEDULER_CHECK);
-                context.startService(scheduler);
-            }
-        }
-    }
-
     /**
      * Received AWARE broadcasts
      * - ACTION_AWARE_SYNC_DATA
@@ -500,6 +475,7 @@ public class Applications extends AccessibilityService {
      * @author df
      */
     private static final Applications_Broadcaster awareMonitor = new Applications_Broadcaster();
+
     public static class Applications_Broadcaster extends WakefulBroadcastReceiver {
         @Override
         public void onReceive(Context context, Intent intent) {
